Incorrect permissions raise a 404 instead of admin login prompt.

+++ apps/hypervisor/views.py
@@ -12,8 +12,9 @@ from django.contrib import messages
 import persistent_messages
 import simplejson
 
-@staff_member_required
 def view(request, pk):
+  if not request.user.is_staff: raise Http404
+
   instance = get_object_or_404(Hypervisor, pk=pk)
 
   storagepools = StoragePool.objects.filter(hypervisor=instance)
@@ -56,16 +57,16 @@ def view(request, pk):
     },
     context_instance=RequestContext(request))
 
-@staff_member_required
 def index(request):
+  if not request.user.is_staff: raise Http404
   hypervisors = Hypervisor.objects.all()
   return render_to_response('hypervisor/index.html', {
       'hypervisors': hypervisors,
     },
     context_instance=RequestContext(request))
 
-@staff_member_required
 def add(request):
+  if not request.user.is_staff: raise Http404
   form = HypervisorForm()
   
   if request.method == "POST":
@@ -91,8 +92,8 @@ def add(request):
     },
     context_instance=RequestContext(request))
 
-@staff_member_required
 def edit(request):
+  if not request.user.is_staff: raise Http404
   if request.is_ajax() and request.method == 'POST':
     json = request.POST
     try:
@@ -138,30 +139,30 @@ def edit(request):
     return HttpResponse('{}', mimetype="application/json")
   raise Http404
 
-@staff_member_required
 def start(request, pk):
+  if not request.user.is_staff: raise Http404
   hypervisor = get_object_or_404(Hypervisor, pk=pk)
   hypervisor.start()
   messages.add_message(request, persistent_messages.SUCCESS,
     'Started Hypervisor %s' % (hypervisor))
   return redirect('/hypervisor/')
   
-@staff_member_required
 def stop(request, pk):
+  if not request.user.is_staff: raise Http404
   hypervisor = get_object_or_404(Hypervisor, pk=pk)
   hypervisor.stop()
   messages.add_message(request, persistent_messages.SUCCESS,
     'Stopped Hypervisor %s' % (hypervisor))
   return redirect('/hypervisor/')
 
-@staff_member_required
 def delete(request, pk):
+  if not request.user.is_staff: raise Http404
   hypervisor = get_object_or_404(Hypervisor, pk=pk)
   hypervisor.delete()
   return redirect('/hypervisor/')
 
-@staff_member_required
 def update(request, pk):
+  if not request.user.is_staff: raise Http404
   hypervisor = get_object_or_404(Hypervisor, pk=pk)
   conn = hypervisor.get_connection(True)
   return redirect('/hypervisor/')
